update - truncate probability

+++ main.py
@@ -1,13 +1,14 @@
 #!/bin/python 
 
-from flask import Flask, render_template, request, flash, redirect, url_for, make_response
+from flask import Flask, render_template, request, redirect, url_for, make_response
 from werkzeug.utils import secure_filename
 from werkzeug.datastructures import  FileStorage
 from fastai.vision.all import *
 import os
 import time
 import datetime
-
+from pathlib import Path
+import re
 
 app = Flask(__name__)
 app.config['MAX_CONTENT_LENGTH'] = 1024 * 1024
@@ -17,30 +18,25 @@ app.config['UPLOAD_FOLDER'] = 'static'
 
 model = load_learner('result-resnet34.pkl')
 
+def truncate(num):
+    return re.sub(r'^(\d+\.\d{,3})\d*$',r'\1',str(num))
+
 # remove files older than 1 minute (60s)
 def remove_files():
-   for file in [os.path.join(app.config['UPLOAD_FOLDER'],file)
-            for file in next(os.walk(app.config['UPLOAD_FOLDER']))[2]
-            if not file.startswith('.')]:
-
-      file_path = os.path.join(file)
-      creation_date = datetime.datetime.fromtimestamp(os.path.getmtime(file_path))
+   for f in [x for x in Path(app.config['UPLOAD_FOLDER']).iterdir() if x.is_file() and not x.name.startswith('.')]:
+      creation_date = datetime.datetime.fromtimestamp(os.path.getmtime(f))
       now = datetime.datetime.now()
       diff = now - creation_date
       if diff.seconds > 60:
-         print(f'Removing file {file}')
-         os.remove(file_path)
-
-
-def validate_filename(filename):
-   return '.' in filename and filename.rsplit('.', 1)[1].lower() in app.config['UPLOAD_EXTENSIONS']
+         print(f'Removing file {f.name}')
+         os.remove(f)
 
 
 def predict(filename):
    try:
       prediction = model.predict(filename)
       num = int(prediction[1].numpy().tolist())
-      prob = float(prediction[2].numpy()[num])
+      prob = truncate(float(prediction[2].numpy()[num]))
       print(f'Classified as {prediction[0]}, Class number {num} with probability {prob}')
       return {'predicted': prediction[0], 'class_number':num, 'probability': prob}
    except:
@@ -49,15 +45,15 @@ def predict(filename):
 
 def save_file(f):
    
-   remove_files()
-
    filename = secure_filename(f.filename)
    
    if filename=='':
       return (False, 'Empty filename')
 
-   if not validate_filename(filename):
-      return (False, 'The file extension should be .jpg, .png or jpeg')
+   if not filename.lower().endswith(tuple(app.config['UPLOAD_EXTENSIONS'])):
+      return (False, f"The file extension should be {app.config['UPLOAD_EXTENSIONS']}")
+
+   remove_files()
 
    timestamp = time.time()  
    fullname = os.path.join(app.config['UPLOAD_FOLDER'], filename + f'_{timestamp}')
@@ -66,22 +62,19 @@ def save_file(f):
    return (True, fullname)
    
 
-@app.route('/')
-def home():
-    return render_template('upload.html')
-
 @app.route('/check')
 def check():
     return 'ok'
 
 
-@app.route('/upload', methods = ['GET','POST'])
+@app.route('/', methods = ['GET','POST'])
 def upload():
    response = render_template('upload.html')
    filename = ''
-   if request.method != 'GET':
+   if request.method == 'POST':
       f = request.files['file']
       (result, filename) = save_file(f)
+      print(response, filename)
       if result:
          prediction = predict(filename)
          if 'predicted' in prediction.keys():
@@ -89,19 +82,27 @@ def upload():
                filename=filename.split('/')[1], 
                predict_message=f"Classified as {prediction['predicted']} with probability {prediction['probability']}")
          else:
-            flash('An error has occurred')
+            response = render_template('upload.html', error_message='An error has occurred processing file')
       else:
-         flash(filename)
-
+         response = render_template('upload.html', error_message=filename)
    return response
 
 @app.route('/classify', methods = ['POST'])
 def upload_file():
    f = request.files['file']
    (result, filename) = save_file(f)
-   result = predict(filename)
-   os.remove(filename)
-   return result
+   if result:
+      result = predict(filename)
+      os.remove(filename)
+      return result
+   else:
+      return filename
+
+@app.errorhandler(Exception)          
+def basic_error(e):
+   print('Error:', e)          
+   return render_template('upload.html', error_message='An error has occurred')         
+  
 
 if __name__ == '__main__':
     app.run(debug=True, host='0.0.0.0', port=5000)
